requirements.txt
django-haystack>=2.0.0
elasticsearch==1.4.0
